Escape Analysis
* Uses intra-method and inter-method analysis
to identify escape state for objects.
* It uses connection graphs to track escape state,
and escape relationships using points-to,
deferred and field links.
* Escape state values are no escape,
argument escape or global escape.
* Preemptively assumes that all exceptions thrown
propagate to other methods,
hence mark them as argument escaping.
* Dot files for donnection graphs generated
for both intra-method and inter method phases.
* This enables to see the effects of each analysis phase,
which will help debugging.
* Make Runnable instances default to global escape.
* Include self when iterating over points to updating caller
This is to capture ParameterValue instances
that are passed in, but not really accessed,
beyond say, being set to a global variable.
* Handle interface calls in EA,
both for generic and non-generic versions.
* Interface implementations are visited before a union.
* Support casting of variables.
* Support merging CGs of non-abstract subclasses.
* Also handle abstract methods.
* Make sure New in optimize is ordered correctly,
as per the original dependency.
* Zero stack allocated objects in EA using helper

Fix EA for latest main changes

Update EA for latest main changes